Users(admin, doctors, patients and intermediate people) can be searched via their ids.

+++ healthcare/views.py
@@ -133,45 +133,57 @@ def doctors(request):
         
 @api_view(['GET', 'POST', 'PUT', 'PATCH', 'DELETE'])
 def new_Users(request):
-    if request.method == 'GET':
-        patients = NewUser.objects.all()
-        serializer = NewUserSerializer(patients, many=True)
-        return Response(serializer.data)
-    
-    elif request.method == 'POST':
-        data = request.data
-        serializer = NewUserSerializer(data=data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors)
-    
-    elif request.method == 'PUT':
-        data = request.data
-        serializer = NewUserSerializer(data=data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors)
-    
-    elif request.method == 'PATCH':
-        data = request.data
-        obj = NewUser.objects.get(id=data['id'])
-        serializer = NewUserSerializer(obj, data=data, partial=True)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors)
-    
-    elif request.method == 'DELETE':
-        data = request.data
+    user_id = request.query_params.get('id')
+    if user_id is not None:
         try:
-            obj = NewUser.objects.get(id=data['id'])
-            name = obj.username
-            obj.delete()
-            return Response({'message' : f"{name} deleted successfully."})
+            new_User = NewUser.objects.filter(id=user_id)
+            serializer = NewUserSerializer(new_User)
+            return Response(serializer.data)
         except Exception as e:
-            return Response({'message' : f"Error is {e}"})
+            return Response({
+                'status' : 404,
+                'message' : f"Error : {e}"
+            })
+    else:
+        if request.method == 'GET':
+            patients = NewUser.objects.all()
+            serializer = NewUserSerializer(patients, many=True)
+            return Response(serializer.data)
+
+        elif request.method == 'POST':
+            data = request.data
+            serializer = NewUserSerializer(data=data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            return Response(serializer.errors)
+    
+        elif request.method == 'PUT':
+            data = request.data
+            serializer = NewUserSerializer(data=data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            return Response(serializer.errors)
+    
+        elif request.method == 'PATCH':
+            data = request.data
+            obj = NewUser.objects.get(id=data['id'])
+            serializer = NewUserSerializer(obj, data=data, partial=True)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            return Response(serializer.errors)
+    
+        elif request.method == 'DELETE':
+            data = request.data
+            try:
+                obj = NewUser.objects.get(id=data['id'])
+                name = obj.username
+                obj.delete()
+                return Response({'message' : f"{name} deleted successfully."})
+            except Exception as e:
+                return Response({'message' : f"Error is {e}"})
     
 @api_view(['GET'])
 def home(request):
